+++ team_integration_test.go
@@ -314,6 +314,7 @@ func TestTeam_Unmarshal(t *testing.T) {
 					"manage-policies":     true,
 					"manage-workspaces":   true,
 					"manage-vcs-settings": true,
+					"manage-projects":     true,
 				},
 				"permissions": map[string]interface{}{
 					"can-destroy":           true,
@@ -330,13 +331,13 @@ func TestTeam_Unmarshal(t *testing.T) {
 	err = unmarshalResponse(responseBody, team)
 	require.NoError(t, err)
 
-	require.NoError(t, err)
 	assert.Equal(t, team.ID, "1")
 	assert.Equal(t, team.Name, "team hashi")
 	assert.Empty(t, team.SSOTeamID)
 	assert.Equal(t, team.OrganizationAccess.ManageWorkspaces, true)
 	assert.Equal(t, team.OrganizationAccess.ManageVCSSettings, true)
 	assert.Equal(t, team.OrganizationAccess.ManagePolicies, true)
+	assert.Equal(t, team.OrganizationAccess.ManageProjects, true)
 	assert.Equal(t, team.Permissions.CanDestroy, true)
 	assert.Equal(t, team.Permissions.CanUpdateMembership, true)
 }
@@ -399,6 +400,50 @@ func TestTeamsUpdateRunTasks(t *testing.T) {
 			assert.Equal(t,
 				*options.OrganizationAccess.ManageRunTasks,
 				item.OrganizationAccess.ManageRunTasks,
+			)
+		}
+	})
+}
+
+func TestTeamsUpdateManageProjects(t *testing.T) {
+	skipIfFreeOnly(t)
+	skipUnlessBeta(t)
+	skipIfEnterprise(t)
+
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	defer orgTestCleanup()
+
+	tmTest, tmTestCleanup := createTeam(t, client, orgTest)
+	defer tmTestCleanup()
+
+	t.Run("with valid options", func(t *testing.T) {
+		options := TeamUpdateOptions{
+			Name: String("foo bar"),
+			OrganizationAccess: &OrganizationAccessOptions{
+				// **Note: ManageProjects requires ManageWorkspaces field to be set and subject to change later.**
+				ManageWorkspaces: Bool(true),
+				ManageProjects:   Bool(true),
+			},
+			Visibility: String("organization"),
+		}
+
+		tm, err := client.Teams.Update(ctx, tmTest.ID, options)
+		require.NoError(t, err)
+
+		refreshed, err := client.Teams.Read(ctx, tmTest.ID)
+		require.NoError(t, err)
+
+		for _, item := range []*Team{
+			tm,
+			refreshed,
+		} {
+			assert.Equal(t, *options.Name, item.Name)
+			assert.Equal(t,
+				*options.OrganizationAccess.ManageProjects,
+				item.OrganizationAccess.ManageProjects,
 			)
 		}
 	})